switched to action as dal is for fetching
